requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bottleneck=1.3.5=py39h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotli-python=1.0.9=py39h6a678d5_7
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.12.12=h06a4308_0
certifi=2023.11.17=py39h06a4308_0
cffi=1.16.0=py39h5eee18b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
config=0.5.1=pypi_0
contourpy=1.2.0=py39hdb19cb5_0
cryptography=41.0.3=py39hdda0065_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
diffusers=0.12.1=pypi_0
einops=0.8.0=pypi_0
et_xmlfile=1.1.0=py39h06a4308_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.13.1=py39h06a4308_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.3.1=pypi_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h5py=3.9.0=py39he06866b_0
hdf5=1.12.1=h2b7332f_3
huggingface-hub=0.22.2=pypi_0
icu=73.1=h6a678d5_0
idna=3.4=py39h06a4308_0
importlib-metadata=7.1.0=pypi_0
importlib_resources=6.1.0=py39h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.2=py39h06a4308_0
joblib=1.2.0=py39h06a4308_0
jpeg=9e=h5eee18b_1
kiwisolver=1.4.4=py39h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.7.2.10=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.3.141=0
libcurl=8.5.0=h251f7ec_0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libidn2=2.3.4=h5eee18b_0
libllvm14=14.0.6=hdb19cb5_3
libnghttp2=1.57.0=h2d74bed_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libpq=12.15=hdbd6064_1
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hf1b16e4_1
loguru=0.5.3=py39h06a4308_4
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib=3.7.2=py39h06a4308_0
matplotlib-base=3.7.2=py39h1128e8f_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpi=1.0=mpich
mpi4py=3.1.4=py39hfc96bbd_0
mpich=3.3.2=external_0
mpmath=1.3.0=py39h06a4308_0
munkres=1.1.4=py_0
mypy-extensions=1.0.0=pypi_0
mysql=5.7.24=h721c034_2
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py39h06a4308_0
numexpr=2.8.7=py39h85018f9_0
numpy=1.26.0=py39h5f9d8c6_0
numpy-base=1.26.0=py39hb5e798b_0
opencv-python=4.9.0.80=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openpyxl=3.0.10=py39h5eee18b_0
openssl=3.0.12=h7f8727e_0
packaging=23.1=py39h06a4308_0
pandas=2.0.3=py39h1128e8f_0
pcre=8.45=h295c915_0
pillow=10.0.1=py39ha6cbd5a_0
pip=23.3=py39h06a4308_0
ply=3.11=py39h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py39h06a4308_0
pyparsing=3.0.9=py39h06a4308_0
pyqt=5.15.10=py39h6a678d5_0
pyqt5-sip=12.13.0=py39h5eee18b_0
pyrallis=0.3.1=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.18=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.0.1=py3.9_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py39h06a4308_0
pyyaml=6.0.1=pypi_0
qt-main=5.15.2=h53bd1ea_10
readline=8.2=h5eee18b_0
regex=2023.10.3=pypi_0
requests=2.27.1=pypi_0
sacremoses=0.1.1=pypi_0
scikit-learn=1.3.0=py39h1128e8f_0
scipy=1.11.3=py39h5f9d8c6_0
seaborn=0.12.2=py39h06a4308_0
setuptools=68.0.0=py39h06a4308_0
sip=6.7.12=py39h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
sympy=1.11.1=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
tomli=2.0.1=py39h06a4308_0
torchaudio=2.0.2=py39_cu117
torchtriton=2.0.0=py39
torchvision=0.15.2=py39_cu117
tornado=6.3.3=py39h5eee18b_0
tqdm=4.66.1=pypi_0
transformers=4.26.0=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.7.1=py39h06a4308_0
tzdata=2023c=h04d1e81_0
ujson=5.4.0=py39h6a678d5_0
urllib3=1.26.18=py39h06a4308_0
vit-pytorch=1.8.5=pypi_0
wheel=0.41.2=py39h06a4308_0
x264=1!157.20191217=h7b6447c_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zipp=3.11.0=py39h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
